Required modifications and implementation comparison
